Added logg in TMS,Restructured Auth
